requirements-dev.txt
audata==1.0.3
backports.zoneinfo==0.2.1
cycler==0.11.0
EditorConfig==0.12.3
fonttools==4.31.2
h5py==2.10.0
heartpy==1.2.7
joblib==1.1.0
jsbeautifier==1.14.2
kiwisolver==1.4.2
lorem==0.1.1
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tqdm==4.63.1
tzdata==2022.1
tzlocal==4.1
